requirements.txt
pandas==2.0.3
numpy==1.24.3
pymysql==1.1.0
sqlalchemy==2.0.20
scikit-learn==1.3.0
xgboost==2.0.0
catboost==1.2
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
joblib==1.3.2
flask==2.3.3
flask-cors==4.0.0